update node/browser webpack logic
