+++ lib/crypto/base64.ts
@@ -10,8 +10,7 @@
  * See the License for the specific language governing permissions and limitations under the License.
  */
 
-import atob from './atob';
-import btoa from './btoa';
+import { atob, btoa } from './webcrypto';
 
 // converts a string to base64 (url/filename safe variant)
 export function stringToBase64Url(str) {
