+++ lib/crypto/browser.ts
@@ -1,0 +1,6 @@
+/* global atob, btoa, crypto */
+const a = atob;
+const b = btoa;
+const c = crypto;
+
+export { a as atob, b as btoa, c as webcrypto };
