+++ lib/crypto/index.ts
@@ -1,9 +1,4 @@
-import atob from './atob';
-import btoa from './btoa';
-import webcrypto from './webcrypto';
-
-export { atob, btoa, webcrypto };
-
 export * from './base64';
 export * from './oidcHash';
 export * from './verifyToken';
+export * from './webcrypto';
