+++ lib/idx/authenticate.ts
@@ -1,84 +1,9 @@
 /* eslint-disable complexity */
-import { AuthApiError } from '../errors';
 import { AuthTransaction } from '../tx';
-import { OktaAuth, AuthorizeOptions, IdxResponse, isRawIdxResponse, APIError, IdxTransactionMeta } from '../types';
+import { OktaAuth, AuthorizeOptions, IdxTransactionMeta } from '../types';
 import { interact } from './interact';
 import { introspect } from './introspect';
-
-function createApiError(res): APIError {
-  console.log('CREATE API ERROR: ', JSON.stringify(res.messages, null, 2));
-  let allErrors = [];
-
-  if (res.messages && Array.isArray(res.messages.value)) {
-    allErrors = res.messages.value.map(o => o.message);
-  }
-
-  return new AuthApiError({
-    errorSummary: allErrors.join('. '),
-    errorCauses: allErrors
-  });
-}
-
-function canSatisfyRemediation(options: AuthorizeOptions, idxResponse: IdxResponse) {
-  const { username, password } = options;
-  const { neededToProceed } = idxResponse;
-  if (neededToProceed.length !== 1) {
-    return false;
-  }
-
-  const form = neededToProceed[0];
-  if (form.name === 'identify') {
-    if (username && form.value.find(el => el.name === 'identifier')) {
-      return true;
-    }
-    if (password && form.value.find(el => el.name === 'credentials')) {
-      return true;
-    }
-  }
-
-  if (form.name === 'challenge-authenticator') {
-    if (password && form.value.find(el => el.name === 'credentials')) {
-      return true;
-    }
-  }
-
-  return false;
-}
-
-async function satisfyRemediations(
-  authClient: OktaAuth,
-  options: AuthorizeOptions,
-  idxResponse: IdxResponse,
-  stateHandle: string
-) {
-  // Recursive loop breaker
-  if (!canSatisfyRemediation(options, idxResponse)) {
-    console.log('REMEDIATION CANNOT BE SATISIFIED', idxResponse);
-    return idxResponse;
-  }
-
-  const { username, password } = options;
-  const { neededToProceed } = idxResponse;
-
-  const form = neededToProceed[0];
-  const data = { stateHandle, identifier: undefined, credentials: undefined };
-  if (form.value.find(el => el.name === 'identifier')) {
-    data.identifier = username;
-  }
-  if (form.value.find(el => el.name === 'credentials')) {
-    data.credentials = { passcode: password };
-  }
-  console.log('PASSING DATA to proceed: ', data);
-  try {
-    idxResponse = await idxResponse.proceed(form.name, data);
-    return satisfyRemediations(authClient, options, idxResponse, stateHandle); // recursive call
-  } catch (e) {
-    if (isRawIdxResponse(e)) { // idx responses are sometimes thrown
-      throw createApiError(e);
-    }
-    throw e;
-  }
-}
+import { remediate } from './remediate';
 
 export async function authenticate(authClient: OktaAuth, options: AuthorizeOptions): Promise<AuthTransaction> {
   let { state, scopes, interactionHandle } = options;
@@ -97,7 +22,7 @@ export async function authenticate(authClient: OktaAuth, options: AuthorizeOptio
   let idxResponse = await introspect(authClient, { stateHandle });
 
   // Can we handle the remediations?
-  idxResponse = await satisfyRemediations(authClient, options, idxResponse, stateHandle);
+  idxResponse = await remediate(authClient, options, idxResponse, stateHandle);
 
   // Did we get an interaction code?
   let status = 'FAILURE';
